Move get_config
